Update port layer (#1966)
* Add Focus, Blur function into LWE::WebView.
* Revise some platform-defines
* Revise CharacterData functions. they are accept offset, length as UTF-16

+++ src/core/util/String.cpp
@@ -1381,7 +1381,7 @@ std::string StringUtils::toBase64HTMLDataURI(const std::string& src,
     return dataURI;
 }
 
-int utf32ToUtf16(char32_t i, char16_t* u)
+size_t utf32ToUtf16(char32_t i, char16_t* u)
 {
     if (i <= 0xffff) {
         if (i >= 0xd800 && i <= 0xdfff) {
